Internal: validate options using a function per option
