+++ src/configurator/options.js
@@ -1,0 +1,4 @@
+export appcache from "./options/appcache.js"
+export assets from "./options/assets.js"
+export baseUrl from "./options/baseUrl.js"
+export production from "./options/production.js"
